# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Prime Home LB Ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTS ───────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (user's own or admin all)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── WISHLIST ─────────────────────────────────────────────────────────────────
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidateResult"

  # ─── SETTINGS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        productCount: { type: integer }
        createdAt: { type: string }
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        imageUrl: { type: string }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        imageUrl: { type: string }

    # ─── PRODUCT ─────────────────────────────────────────────────────────────
...
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
...
    OrderItem:
      type: object
      required: [productId, productName, price, quantity]
      properties:
        productId: { type: integer }
        productName: { type: string }
        productImageUrl: { type: ["string", "null"] }
        price: { type: number }
        quantity: { type: integer }
        subtotal: { type: number }
    Order:
      type: object
      required: [id, orderNumber, status, total, createdAt]
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        status: { type: string }
        customerName: { type: string }
        customerEmail: { type: ["string", "null"] }
        customerPhone: { type: string }
        address: { type: string }
        city: { type: string }
        deliveryArea: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        subtotal: { type: number }
        deliveryFee: { type: number }
        discount: { type: number }
        total: { type: number }
        paymentMethod: { type: string }
        couponCode: { type: ["string", "null"] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt: { type: string }
    OrderList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    OrderInput:
      type: object
      required: [customerName, customerPhone, address, city, paymentMethod, items]
      properties:
        customerName: { type: string }
        customerEmail: { type: string }
        customerPhone: { type: string }
        address: { type: string }
        city: { type: string }
        deliveryArea: { type: string }
        notes: { type: string }
        paymentMethod: { type: string }
        couponCode: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemInput"
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }

    # ─── WISHLIST ─────────────────────────────────────────────────────────────
...
    Coupon:
      type: object
      required: [id, code, type, value, active]
      properties:
        id: { type: integer }
        code: { type: string }
        type: { type: string }
        value: { type: number }
        minOrderAmount: { type: ["number", "null"] }
        maxUses: { type: ["integer", "null"] }
        usedCount: { type: integer }
        expiresAt: { type: ["string", "null"] }
        active: { type: boolean }
        createdAt: { type: string }
    CouponInput:
      type: object
      required: [code, type, value]
      properties:
        code: { type: string }
        type: { type: string, enum: [percentage, fixed] }
        value: { type: number }
        minOrderAmount: { type: number }
        maxUses: { type: integer }
        expiresAt: { type: string }
    CouponValidateInput:
      type: object
      required: [code, orderAmount]
      properties:
        code: { type: string }
        orderAmount: { type: number }
    CouponValidateResult:
      type: object
      required: [valid, discount]
      properties:
        valid: { type: boolean }
        discount: { type: number }
        message: { type: ["string", "null"] }

    # ─── SETTINGS ─────────────────────────────────────────────────────────────
...