# lib/api-zod/src/generated/types/couponValidateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prime Home LB Ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CouponValidateResult {
  valid: boolean;
  discount: number;
  /** @nullable */
  message?: string | null;
}