# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prime Home LB Ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  customerName?: string;
  /** @nullable */
  customerEmail?: string | null;
  customerPhone?: string;
  address?: string;
  city?: string;
  /** @nullable */
  deliveryArea?: string | null;
  /** @nullable */
  notes?: string | null;
  subtotal?: number;
  deliveryFee?: number;
  discount?: number;
  total: number;
  paymentMethod?: string;
  /** @nullable */
  couponCode?: string | null;
  items?: OrderItem[];
  createdAt: string;
}